:sparkles: 添加查询员工接口

+++ athena-contract-thrift/src/thrift/service.thrift
@@ -35,7 +35,7 @@ service StaffService {
 
    dto.AthenaResponse addStaff(1:dto.CompanyInfo companyInfo, 2:dto.StaffInfo staffInfo),
 
-   dto.StaffInfo getStaffById(1:dto.CompanyInfo companyInfo, 2:i32 id),
+   dto.StaffInfo getStaffById(1:dto.CompanyInfo companyInfo, 2:i64 id),
 
    list<dto.StaffInfo> getStaffByName(1:dto.CompanyInfo companyInfo, 2:string name),
